core: Shift Skill model from Resume app to CORE app
